requirements.txt
requests>=2.31.0
eth-account>=0.11.0
PyNaCl>=1.5.0
python-dotenv>=1.0.0
base58>=2.1.0
PyJWT>=2.8.0
